beta 0.2.0.1 - support both armv7/arm64 in podspec (pod version >= 1.5.0 required) - refactor neg axis support - fix memory overlap in de-conv - fix CONVOLUTION_TILED_NUMBER spell error - fix few warnings - add binary / interp / permute / relu / reshape / softmax support and optimize conv for OpenGL backend - add clean in nmake build script

+++ source/backend/cpu/compute/CommonOptFunction.cpp
@@ -541,7 +541,7 @@ void MNNRelu6(float* dst, const float* src, size_t size) {
 }
 
 void MNNExp(float* dst, const float* src, size_t dataSize) {
-    int countC8        = dataSize / 8;
+    int countC8        = (int)dataSize / 8;
     if (countC8 > 0) {
         // Align to eight so asm is easier to write
         static float parameters[] = {
